% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{add_stylize_plot_plotting_control}
\alias{add_stylize_plot_plotting_control}
\title{add_stylize_plot_plotting_control}
\usage{
add_stylize_plot_plotting_control()
}
\description{
This function adds UI elements to allow to customize the plot appearance
}
