% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body_composition.R
\name{body_composition}
\alias{body_composition}
\title{body_composition}
\usage{
body_composition(
  finalC1,
  finalC1meta,
  input,
  output,
  session,
  global_data,
  scaleFactor
)
}
\arguments{
\item{finalC1}{input data}

\item{finalC1meta}{combined metadata}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{global_data}{dictionary to store variables session-based for users}

\item{scaleFactor}{used to scale energy expenditure units correctly}
}
\description{
This function generates and overview of body composition and other recorded metadata
}
\examples{
body_composition(values, full_metadata, input, output, session, global_data, 1)
}
