% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_ancova_alternative.R
\name{calculate_statistic}
\alias{calculate_statistic}
\title{calculate_statistic}
\usage{
calculate_statistic(data, method)
}
\description{
This function calculate statistic based on provided method: mean or median
}
