% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{coarsen_data_sets}
\alias{coarsen_data_sets}
\title{coarsen_data_sets}
\usage{
coarsen_data_sets(df, coarsening_factor)
}
\arguments{
\item{df}{data frame}

\item{coarsening_factor}{coarsening factor}
}
\description{
This function coarsens a data frame according to the numerical coarsening factor
}
\examples{
coarsen_data_sets(values, 2)
}
