% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rmr.R
\name{create_df}
\alias{create_df}
\title{create_df}
\usage{
create_df(df, component, M, N, percentage = 1, interval_length = 15)
}
\description{
This function prepares the data frame for resting metabolic rate calculation
}
