% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_model.R
\name{create_lme_model_ui}
\alias{create_lme_model_ui}
\title{create_lme_model_ui}
\usage{
create_lme_model_ui(input, output, true_metadata, df_to_plot, my_dep_var)
}
\description{
This function creates UI for LME modelling
}
