% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{cv}
\alias{cv}
\title{cv}
\usage{
cv(mydf, window = 2)
}
\arguments{
\item{mydf}{data frame}

\item{window}{window size}
}
\description{
This function calculates the coefficient of variation for data frmae
}
