% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_exporter.R
\name{do_export_all_data}
\alias{do_export_all_data}
\title{do_export_all_data}
\usage{
do_export_all_data(
  input,
  output,
  session,
  file_output,
  do_plotting,
  global_data
)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{file_output}{output file path chosen by user}

\item{do_plotting}{plotting routines}

\item{global_data}{hash table containing variables during session}
}
\description{
This function exports all data calculated during the app usage as an archive
}
