% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rmr.R
\name{do_extract}
\alias{do_extract}
\title{do_extract}
\usage{
do_extract(df, component = "O2", percentage = 5, N)
}
\description{
This function does the heavy lifting, calculating the interval-based resting
metabolic rates based on the coefficient of variation criteria
}
