% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{draw_day_night_rectangles}
\alias{draw_day_night_rectangles}
\title{draw_day_night_rectangles}
\usage{
draw_day_night_rectangles(
  df,
  p,
  light_start = 7,
  light_end = 19,
  light_offset = 0,
  day_color = "yellow",
  night_color = "grey",
  light_cycle = c("Day", "Night")
)
}
\arguments{
\item{df}{data frame}

\item{p}{plot object}

\item{light_start}{start of light cycle}

\item{light_end}{end of light cycle}

\item{light_offset}{offset of light cycle from start of day (for zeitgeber time always 0)}

\item{day_color}{color for light phase (day)}

\item{night_color}{color for dark phase (night)}

\item{light_cycle}{choose Day or Night}
}
\description{
This function inks the day with a light color and night with a darker color
}
