% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{enrich_with_metadata}
\alias{enrich_with_metadata}
\title{enrich_with_metadata}
\usage{
enrich_with_metadata(finalC1, C1meta, havemetadata, metadatafile)
}
\arguments{
\item{finalC1}{data frame}

\item{C1meta}{metadata from data files in finalC1}

\item{metadatafile}{structured metadata sheet}

\item{havemetadatafile}{boolean to indicate if we have structured metadata sheet}
}
\description{
Enrich data frame with metadata
}
\examples{
enrich_with_metadata(values, metadata_basic, TRUE, "metadata.xlsx")
}
