% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rmr.R
\name{extract_rmr}
\alias{extract_rmr}
\title{extract_rmr}
\usage{
extract_rmr(data, M = 5, PERCENTAGE = 5, interval_length = 15)
}
\description{
This function extracts the resting metabolic rate from the data frame
}
