% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rmr_helper.R
\name{extract_rmr_helper}
\alias{extract_rmr_helper}
\title{extract_rmr_helper}
\usage{
extract_rmr_helper(interval_length = 15, percentage_best = 1, M = 1)
}
\description{
This function is a helper method/wrapper to calculate the resting metabolic rate
}
