% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{generate_statistical_table}
\alias{generate_statistical_table}
\title{generate_statistical_table}
\usage{
generate_statistical_table(results)
}
\description{
This function generate statistical table in case we have multiple comparisons
}
