% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{get_constants}
\alias{get_constants}
\title{get_constants}
\usage{
get_constants(file)
}
\arguments{
\item{file}{input file}
}
\description{
This function gets constant and value pairs from metadata sheet
}
\examples{
get_constants(input_file)
}
