% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{get_covariates_and_units}
\alias{get_covariates_and_units}
\title{get_covariates_and_units}
\usage{
get_covariates_and_units(file)
}
\arguments{
\item{file}{input_file}
}
\description{
This function gets covariates with units from metadata sheet
}
\examples{
get_covariates_and_units(input_file)
}
