% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_info.R
\name{get_git_information_from_repository}
\alias{get_git_information_from_repository}
\title{get_git_info_from_repository}
\usage{
get_git_information_from_repository()
}
\description{
This function is used to create the git branch information for the UI
}
