% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_new_download_buttons}
\alias{get_new_download_buttons}
\title{get_new_download_buttons}
\usage{
get_new_download_buttons(plot_div = "")
}
\description{
This function generate new download buttons to get plot as specified by div element
}
