% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_ancova_alternative.R
\name{get_r_squared_clean}
\alias{get_r_squared_clean}
\title{get_r_squared_clean}
\usage{
get_r_squared_clean(rvalue)
}
\description{
This function get's the R-squared value in a clean way for plotly
}
