% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{get_study_description_from_metadata}
\alias{get_study_description_from_metadata}
\title{get_study_description_from_metadata}
\usage{
get_study_description_from_metadata(file)
}
\arguments{
\item{file}{input file}
}
\description{
This function extract study description from metadata sheet
}
\examples{
get_study_description_from_metadata(input_file)
}
