% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_time_diff}
\alias{get_time_diff}
\title{get_time_diff}
\usage{
get_time_diff(df, from = 2, to = 3, do_warn = FALSE)
}
\description{
This function get's the time difference (measurement interval length) in minutes
}
