% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{get_true_metadata}
\alias{get_true_metadata}
\title{get_true_metadata

This function loads the true metadata from the Metadata Sheet (.xlsx)}
\usage{
get_true_metadata(file, load_example_data)
}
\arguments{
\item{file}{input file}

\item{load_example_data}{TRUE if example data should be loaded (default: FALSE)}
}
\description{
get_true_metadata

This function loads the true metadata from the Metadata Sheet (.xlsx)
}
\examples{
get_true_metadata(input_file, FALSE)
}
