% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide.R
\docType{data}
\name{guide}
\alias{guide}
\title{guide}
\format{
An object of class \code{Cicerone} (inherits from \code{R6}) of length 15.
}
\usage{
guide
}
\description{
This function creates a simple user guide navigating the user through the app
}
\keyword{datasets}
