% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cosmed_helper.R
\name{import_cosmed}
\alias{import_cosmed}
\title{import_cosmed}
\usage{
import_cosmed(file, file_out)
}
\arguments{
\item{file}{input file}

\item{file_out}{output file}
}
\description{
This function imports a COSMED indirect calorimetry data file
}
\examples{
import_cosmed(input_file, output_file)
}
