% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_pheno_v8_helper.R
\name{import_pheno_v8}
\alias{import_pheno_v8}
\title{import_pheno_v8}
\usage{
import_pheno_v8(file, file_out)
}
\arguments{
\item{file}{input file}

\item{file_out}{output file}
}
\description{
This function imports TSE Systems PhenoMaster v8 indirect calorimetry data sets
}
\examples{
import_pheno_v8(input_file, output_file)
}
