% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_promethion_helper.R
\name{import_promethion}
\alias{import_promethion}
\title{import_promethion}
\usage{
import_promethion(file, file_out)
}
\arguments{
\item{file}{input file}

\item{file_out}{output file}
}
\description{
This function imports Sable Systems Prometion indirect calorimetry data sets
}
\examples{
import_promethion(input_file, output_file)
}
