% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rmr_helper.R
\name{lookup_cohort_belonging}
\alias{lookup_cohort_belonging}
\title{lookup_cohort_belonging}
\usage{
lookup_cohort_belonging(interval_length_list_per_cohort_and_animals, id)
}
\description{
This function looks up the cohort a given animal id belongs to
}
