% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rmr_helper.R
\name{lookup_interval_length}
\alias{lookup_interval_length}
\title{lookup_interval_length}
\usage{
lookup_interval_length(interval_length_list_per_cohort_and_animals, value)
}
\description{
This function is used to lookup the interval length based on a cohort
}
