% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_model.R
\name{model_effects}
\alias{model_effects}
\title{model_effects}
\usage{
model_effects(df, dep_var, input)
}
\arguments{
\item{df}{data frame}

\item{dep_var}{dependent variable}

\item{input}{shiny input}
}
\description{
This function uses a linear mixed effect model to explain effects on a dependent var
}
