% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locomotion.R
\name{plot_locomotion}
\alias{plot_locomotion}
\title{plot_locomotion}
\usage{
plot_locomotion(
  file,
  x_min_food = 2,
  x_max_food = 9.5,
  y_min_food = 4,
  y_max_food = 16,
  x_min_scale = 20.5,
  x_max_scale = 29,
  y_min_scale = 4.2,
  y_max_scale = 9.5,
  x_min_bottle = 20.5,
  x_max_bottle = 28,
  y_min_bottle = 12,
  y_max_bottle = 15.5
)
}
\arguments{
\item{file}{input file}

\item{x_min_food}{min x coordinate of food hamper}

\item{x_max_food}{max x coordinate of food hamper}

\item{y_min_food}{min y coordinate of food hamper}

\item{y_max_food}{max y coordinate of food hamper}

\item{x_min_scale}{min x scale factor}

\item{x_max_scale}{max x scale factor}

\item{y_min_scale}{min y scale factor}

\item{y_max_scale}{max y scale factor}

\item{x_min_bottle}{min x coordinate water bottle}

\item{x_max_bottle}{max x coordinate water bottle}

\item{y_min_bottle}{min y coordinate water bottle}

\item{y_max_bottle}{max y coordinate water bottle}
}
\description{
This function plots the locomotion / probability density map of a standard
Sable Systems Promethion file
}
