% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locomotion_budget.R
\name{plot_locomotion_budget}
\alias{plot_locomotion_budget}
\title{plot_locomotion_budget}
\usage{
plot_locomotion_budget(file)
}
\arguments{
\item{file}{input file}
}
\description{
This function plots the locomotional budget from a Sable Systems Promethion file
}
\examples{
plot_locomotion_budget(input_file)
}
