% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_exporter.R
\name{prepare_data_frame_for_export}
\alias{prepare_data_frame_for_export}
\title{prepare_data_frame_for_export}
\usage{
prepare_data_frame_for_export(df_to_plot, global_data, session)
}
\arguments{
\item{df_to_plot}{data frame}

\item{global_data}{hash table containing variables during session}

\item{session}{shiny session}
}
\description{
This function prepares the data frame with calculated quantities for export
}
