% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_rmr_helper.R
\name{pretty_print_interval_length_list}
\alias{pretty_print_interval_length_list}
\title{pretty_print_interval_length_list}
\usage{
pretty_print_interval_length_list(interval_length_list_per_cohort_and_animals)
}
\description{
This function pretty prints the interval length list with cohort information
}
