% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{process_return_value_for_statistic}
\alias{process_return_value_for_statistic}
\title{process_return_value_for_statistic}
\usage{
process_return_value_for_statistic(value, as_ui = TRUE)
}
\description{
Helper method to return value for statistic after calculation, e.g. pretty
print the digits and round
}
