% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_measurement.R
\name{raw_measurement}
\alias{raw_measurement}
\title{raw_measurement}
\usage{
raw_measurement(
  finalC1,
  finalC1meta,
  input,
  output,
  session,
  global_data,
  scaleFactor
)
}
\arguments{
\item{finalC1}{input data}

\item{finalC1meta}{combined metadata}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{global_data}{dictionary to store variables session-based for users}

\item{scaleFactor}{used to scale energy expenditure units correctly}
}
\description{
This function displays raw measurements
}
\examples{
raw_measurement(values, full_metadata, input, output, session, global_data, 1)
}
