% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_cosmed_helper.R
\name{reformat_time_for_cosmed}
\alias{reformat_time_for_cosmed}
\title{reformat_time_for_cosmed}
\usage{
reformat_time_for_cosmed(time)
}
\arguments{
\item{time}{time}
}
\description{
This function reformats the time from cosmed format to general format
}
