% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{remove_z_score_outliers}
\alias{remove_z_score_outliers}
\title{remove_z_score_outliers}
\usage{
remove_z_score_outliers(df, sd = 1)
}
\arguments{
\item{df}{data frame}

\item{sd}{multiple of the standard deviation (default=1)}
}
\description{
This function removes outliers based on multiples of the standard deviation
}
\examples{
remove_z_score_outliers(values, sd=2)
}
