% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{remove_zero_values}
\alias{remove_zero_values}
\title{remove_zero_values}
\usage{
remove_zero_values(df, eps)
}
\arguments{
\item{df}{data frame}

\item{eps}{epsilon threshold}
}
\description{
This function removes zero measurement values, based on epsilon threshold
}
\examples{
remove_zero_values(values, 0.0001)
}
