% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resting_metabolic_rate.R
\name{resting_metabolic_rate}
\alias{resting_metabolic_rate}
\title{resting_metabolic_rate}
\usage{
resting_metabolic_rate(
  finalC1,
  finalC1meta,
  input,
  output,
  session,
  global_data,
  scaleFactor
)
}
\arguments{
\item{finalC1}{input data}

\item{finalC1meta}{combined metadata}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{global_data}{dictionary to store variables session-based for users}

\item{scaleFactor}{used to scale energy expenditure units correctly}
}
\description{
This function calculates the resting metabolic rate
}
\examples{
resting_metabolic_rate(values, full_metadata, input, output, session, global_data, 1)
}
