% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{style_plot}
\alias{style_plot}
\title{style_plot}
\usage{
style_plot(p, input)
}
\arguments{
\item{p}{plot object}

\item{input}{shiny input}
}
\description{
This function is used to stylize the plotting result in the UI
}
