% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_energy_expenditure.R
\name{total_energy_expenditure}
\alias{total_energy_expenditure}
\title{total_energy_expenditure}
\usage{
total_energy_expenditure(
  finalC1,
  C1meta,
  finalC1meta,
  input,
  output,
  session,
  global_data,
  scaleFactor
)
}
\arguments{
\item{finalC1}{input data}

\item{C1meta}{basic metadata}

\item{finalC1meta}{combined metadata}

\item{input}{shiny input}

\item{output}{shiny output}

\item{session}{shiny session}

\item{global_data}{dictionary to store variables session-based for users}

\item{scaleFactor}{used to scale energy expenditure units correctly}
}
\description{
This function calculates the total energy expenditure
}
\examples{
total_energy_expenditure(values, basic_metadata, full_metadata, input, output, session, global_data, 1)
}
