% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_model.R
\name{visualize_model_effects}
\alias{visualize_model_effects}
\title{visualize_model_effects}
\usage{
visualize_model_effects(df, dep_var, input, output)
}
\description{
This functions visualizes the linear mixed effect modelling results
}
