% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{zeitgeber_zeit}
\alias{zeitgeber_zeit}
\title{zeitgeber_zeit}
\usage{
zeitgeber_zeit(df, light_on)
}
\arguments{
\item{df}{data frame}

\item{light_on}{indicates when day starts (light on typically)}
}
\description{
Converts the time to zeitgeber zeit
}
\examples{
zeitgeber_zeit(values, 8)
}
